Event signup: changes structure
The Event model was starting to have a lot of concerns, this will add an
EventSignup to events in order to handle everything concerned to signup for
events.

EventUser handles the signup for each user, allows to answer a custom question,
add a user_type for priority and a group to connect with Introduction.

Adds CSV export for the ones organizing the event.