Notifications: first implementation
Adds a notification class with attached mailer
The notifications show up in the top right corner for the user.
They can be marked as read, and the counter will reduce.

Each notification is implemented with a partial view, the same partial view is
used to render in the mailer as in the views.